Create CameraEngine and CameraBaseEngine
